created new FilterBar component
